# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "A web service to keep track of your well being and the things you are grateful for, written in Java using the Spring Framework.",
    "version": "1.0.0",
    "title": "GratitudeJournal",
    "termsOfService": "http://swagger.io/terms/",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "tags": [
    {
      "name": "user",
      "description": "Operations about users"
    },
    {
      "name": "journal",
      "description": "Operations about journaling"
    }
  ],
  "paths": {
    "/user/{userName}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get information about a user",
        "description": "",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid userName supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Removes an already existing user",
        "description": "",
        "operationId": "removeUser",
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "User removed successfully"
          },
          "400": {
            "description": "userName malformed"
          },
          "404": {
            "description": "User does not exist"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Creates a new user",
        "description": "",
        "operationId": "createUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Parameters malformed"
          },
          "403": {
            "description": "A user with the same userName already exists"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Update an existing user",
        "description": "Update any User variable except the userName and userId.",
        "operationId": "updateUser",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "Parameters malformed"
          },
          "404": {
            "description": "User does not exist"
          }
        }
      }
    },
    "/journal/{userName}": {
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "Get a list with date:journalEntryId pairs for a user.",
        "description": "",
        "operationId": "getEntries",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "post": {
        "tags": [
          "journal"
        ],
        "summary": "Add a journal entry for a user.",
        "description": "",
        "operationId": "addEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Entry object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JournalEntry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JournalEntry"
            }
          },
          "403": {
            "description": "An entry already exists for today."
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/journal/entry/{journalEntryId}": {
      "get": {
        "tags": [
          "journal"
        ],
        "summary": "Get details about a journal Entry.",
        "description": "",
        "operationId": "getEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "journalEntryId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JournalEntry"
            }
          },
          "404": {
            "description": "Entry not found"
          }
        }
      },
      "put": {
        "tags": [
          "journal"
        ],
        "summary": "Update a journal entry.",
        "description": "",
        "operationId": "updateEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "journalEntryId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Entry object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/JournalEntry"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/JournalEntry"
            }
          },
          "404": {
            "description": "Entry not found"
          }
        }
      },
      "delete": {
        "tags": [
          "journal"
        ],
        "summary": "Delete a journal entry.",
        "description": "",
        "operationId": "deleteEntry",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "journalEntryId",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Entry removed successfully"
          },
          "404": {
            "description": "Entry not found"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "integer",
          "format": "int64",
          "minimum": 0,
          "maximum": 6
        },
        "userName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      }
    },
    "NewUser": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        }
      },
      "required": [
        "userName",
        "firstName",
        "lastName"
      ]
    },
    "JournalEntry": {
      "type": "object",
      "properties": {
        "wellBeing": {
          "type": "integer",
          "format": "int64"
        },
        "gratefullForToday": {
          "type": "string"
        },
        "gratefullForTodayDescription": {
          "type": "string"
        },
        "gratefullForInLife": {
          "type": "string"
        },
        "gratefullForInLifeDescription": {
          "type": "string"
        }
      },
      "required": []
    }
  }
}